Moved stub tests into a more readable and maintainable format
